OMAP: UART : Fixes the boot failure for lower baud rates
This patch fixes the boot up of Linux at lower baud rates. Due to
timing issues between the platform device and probe of the driver
getting called, the inactivity timers were getting called even
when the Uart Port was not initialized. This results in crash
at boot up. A null check would fix this issue as this would allow
the timer to come up and exit gracefully.

Change-Id: I62499e51d6f420af0517d2e2d3bc0efb8cebb518
Signed-off-by: Deepak K <deepak.k@ti.com>