ARM: OMAP4+: trim_quirks: Added trim override logic
LPDDR and Smart IO is overrided to the unique recomended value LPDDR
and Smart IO for all OMAP4 silicon revs to avoid instabilities.

The DCC is disabled for OMAP4460, so to lock the DPLL @ 2.4
GHz DPLL trim-override needs to be used.

Trim override logic is used in suspend/resume path in
addition to init, by doing all the necessary checks at boot.

Also implemented workaround for OMAP4430 erratum i684.

Change-Id: I517b2c6e1ab407392e2248f659476a79630319be
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>
[nm@ti.com: cleanups for handling it without hacking in the code
into pm and voltage layer as these are specific quirk handling
for ldos]
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Colin Cross <ccross@android.com>
Signed-off-by: K.Shlyakhovoy <x0155534@ti.com>
[ivan.khoronzhuk@ti.com: ported from p-android-omap-3.0]
Signed-off-by: Ivan Khoronzhuk <ivan.khoronzhuk@ti.com>