OMAP4: USBTLL: Off mode disconnect workaround (Errata i754)
When resuming from OFF mode, the EHCI controller was performing a
reset resume for ports using the TLL interface because the port
status was being cleared. This was caused by the host controller
not restoring the state of the ID pin in the TLL module correctly.
This issue is workarounded by enabling the ID pin pull-up in the TLL
(because the EHCI controller is always the host) and disabling ID
pin events as these are not required for the TLL connection.

[Ruslan]: ported from k3.0 (commit-id fa25183)

Change-Id: Ibc22d77606e05e4dd138819079028ebcf254039e
Signed-off-by: Jon Hunter <jon-hunter@ti.com>
Signed-off-by: Ruslan Bilovol <ruslan.bilovol@ti.com>