mfd: Add omap-usbhs runtime PM support
The usbhs core driver does not enable/disable the intefrace and
fucntional clocks; These clocks are handled by hwmod and runtime
pm, hence insted of the clock enable/disable, the runtime pm
APIS are used. however,the port clocks and tll clocks are
handled by the usbhs core.

In this architecture, the usbhs is core driver and it is parent of
ehci and ohci drivers.The earlier APIs omap_usbhs_enable and
omap_usbhs_disable of usbhs core driver are removed; The ehci and
ohci drivers call the pm_runtime_get_sync and pm_runtime_put_sync
of parent device usbhs core. In fact, the runtime framework takes
care of the get sync and put sync of the child in turn call the
get sync and put sync of parent too; but calling get sync and put
sync of parent is by ASYNC mode; This mode queues the work item
in runtime pm work queue, which not getting scheduled in case of
global suspend path. so, here ehci and ohci runtim pm is not
required, these drivers will pass the parent usbhs dev pointer
in runtime pm APIs.

Signed-off-by: Keshava Munegowda <keshava_mgowda@ti.com>
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>