drivers:misc: ti-st: avoid a misleading dbg msg
Previously the private data of each protocol registered to use ST was
used to determine whether the protocol was registered to use shared
transport or otherwise.
However, now a flag is_registered is maintained to identify whether a
protocol intends to use ST.
Upon closing of the UART the error message relevant to this lack of
un-registration was misleading and this patch fixes that.

Change-Id: I1a452dccd7b4bd53cce923df5b61ce077b7aa466
Signed-off-by: Pavan Savoy <pavan_savoy@ti.com>