OMAP4: HACK: Clear the Keypad module interrupt status
for some reason, if a key is pressed and module
wakeup is enabled, CORE and MPU dont enter low
power RET. Also, when we resume from suspend
there is a pending interrupt that is not cleared,
so we are writing 0x0f in KBD_IRQSTATUS to clear any
Keypad interrupts.
This patch also Enables IO Daisy for KEYPAD. Without this
Keypad can not wakeup device from low power state.

Change-Id: I4efe3a51d3d7b67fe31440aed7f10c9daa2c1902
Signed-off-by: Rajeev Kulkarni <rajeevk@ti.com>