ASoC: sdp4430: avoid any delay on HS/HF power down time
If the switch between HS->HF->HS is less than the power down time
the HS DAC is not power off then switch back to HS.
This is produce a pop due to the open HS with already DAC on.
To avoid this pop the power down time needs to be set to 0.

Change-Id: Iffdcd5a150661742bdfbb0170723d91de004642d
Signed-off-by: Francois Mazard <f-mazard@ti.com>
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>