omap: add a new rproc_user misc driver
A small minimal driver, rproc_user, is added that provides interfaces
to the user-space to be able to start/stop the secure mode.

The current implementation provides only the .read file operation, with
the .write file operation stubbed out. The .write always returns an
error since the entire secure playback infrastructure is not supported
yet.

Change-Id: I84f4b0c6cad38ce1dd4f4125271d7b012ba6002c
Signed-off-by: Suman Anna <s-anna@ti.com>