OMAP:DSS2: Add state notifiers before panel is disabled
Wrapping dssdev->driver->disable, so that we can notify
clients before a panel is disconnected.  Clients need to
be notified so they can cancel any operations that would
be blocked on display VSYNC events.

The notifiers can be extended to handle other state
change notifications in the future if needed by clients.

Change-Id: Ia3140514d6d77c8db1c0829a85b4f804223d6964
Signed-off-by: Lajos Molnar <molnar@ti.com>