OMAP4: HS: USE PPA API to program the POR register instead of Monitor API.
On HS/EMU devices, PPA API needs to be used to set L2X0 POR register
and same is done using a monitor API on GP devices.

The K3.0 code isn't using the PPA API to set the POR on HS/EMU
devices.

Fix the same so that on OMAP4460 ES1.1 where POR can be enabled,
the correct value is programmed.

Change-Id: Idc5914a1946cd6950919cd308f902a5d5b0e2f36
Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
Signed-off-by: Rajeev Kulkarni <rajeevk@ti.com>