serial: omap: uart: allow fine grained control of rts line
Implement correctly the Auto-RTS uart feature.
RTS line should be kept pulled high, when uart is not in use.

Provide infrastructure to pass the RTS line pad mux from board file.

Also RTS control has to be very fine granined as immediately after
RTS goes low, other end can start sending the data.

So make RTS low only when we know uart is ready to handle RHR/Rx interrupts.
Run-time-pm callbacks are the right place to have this RTS control implemented.

System-wide-suspend is a special case that needs to be handled, as the whole
setup gets repeated for tty layer where request_irq() happens again.
So set_termios is the right place if we are coming out of a system-wide suspend,
to control RTS line for flow control.

Change-Id: If05f3fd90c655597fddaf45ee0a9c25f66c9546d
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>