serial: omap: Disable wakeup events after port shutdown
When the port is shutdown we disable interrupts from the port
makes no sense to keep wakeup events enabled for the port
so disable wakeup events also for the port.

Interrupts and wakeup events will be configured back on port open
from startup function.

In uart_shutdown scenarios we have seen without this
patch uart is struck in transistion.

Change-Id: I1547b22ec7155485891e3910f814a27e4acb58ae
Signed-off-by: Govindraj.R <govindraj.raja@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>
[<oleksandr.tyshchenko@ti.com> - forward port from K3.0]
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>