omap4: sdp/panda/tablet: reserve memory for dsp
DSP is configured to use 4 MB of memory. Memory for DSP is
reserved during board initialization. The memory is just
below the reserved memory for all ION heaps. This memory is
reserved at boot-time using the memblock API and it is
reserved only if CONFIG_REMOTE_PROC_DSP is enabled.

Change-Id: I437fb5e2d8c2fd5f900bbaeb213dbe662ac8208c
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>