tty: serial: omap-serial: fix irq handling
There are a few problems in serial_omap_irq:

The interrupt type from IIR register is read incorrectly.  UART
interrupts are an integer value masked out of the IIR register,
not a bitmask like any sensible hardware would use.  Convert
the mask operations to mask-and-compare.  Masking iir with
UART_IIR_RLSI was previously matching RDI, RLSI, RX timeout,
and (incorrectly) TX interrupts.  Expand the original mask test
to testing for each of the 3 RX interrupts.

If the interrupt type is UART_IIR_THRI, but the LSR THRE bit
is not set, the interrupt handler doesn't do anything to clear
the THRI interrupt.  Return IRQ_NONE in this case, so the irq
core can detect an interrupt storm and disable the interrupt
if necessary.

Change-Id: Ic2810e2178fb7e1e7b7786ccbc500da8d4feba97
Signed-off-by: Colin Cross <ccross@android.com>