HACK: fix ehci suspend/resume failure
Its observed with some PHY, the 60Mhz clock gets
cut too soon for OMAP EHCI, leaving OMAP-EHCI in a bad state.

So on starting port suspend, make sure the 60Mhz clock to EHCI
is kept alive using an internal clock, so that EHCi can cleanly
transition its hw state machine on a port suspend.

Its not proven if this is the issue hit on USB3333,
but the symptoms look very similar.

NOTE:
This is core linux file. The right fix is to re-implement the function
for ehci-omap and put the fix there.

Change-Id: I6f9bd7dbc2c228d882a626a8d41541b8a64d5d8a
Signed-off-by: Anand Gadiyar <gadiyar@ti.com>
Signed-off-by: Vikram Pandita <vikram.pandita@ti.com>