Regulator: twl6030: setup regulator state for sleep state
TWL6030 enters into "sleep mode" when the PREQ lines are de-asserted.
This is the condition we enter when we are in APP processor suspend
state. However, TWL6030 regulator defaults for Sleep are set to
auto mode, to be certain that we consume lowest amount of power
by ensuring that the regulators and resources remain disabled when
we achieve sleep state in TWL6030 during suspend, we
explicitly setup the state as disabled for sleep when the regulator
is disabled and set as auto(default) when it is enabled. This
allows the current drivers which may use setmode in conjunction
with enable and disable to continue to get the benefit of the
auto mode.

Based on a patch by Baek.Kyung-Han <wildtaz.baek@samsung.com>

Change-Id: I464c45c87ae469bf0d8e44c6f9bd9e7f38d90a3e
Signed-off-by: Nishanth Menon <nm@ti.com>
[oleksandr.tyshchenko@ti.com: ported from K3.0]
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>