OMAP: HSI: Fix missed interrupt in polling mode
Wen a RX interrupt occurs, it is disabled unconditionnaly whether there is still some polling
ongoing or not. But polling is not one-shot mode, but must remain enabled once set.
Consequently it can happen that despite polling is enabled and HSI driver expects an interrupt to
inform about RX data reception, this interrupt will never arrive more than once.

This issue was fixed in K35, but didn't make it through K3.0

Change-Id: I62ad87ed54eff4022260f0f4c781cd74996f49f3
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>