blaze_defconfig: increase vmalloc space to 768MB
The default vmalloc space of 128MB results in 768MB of memory
used as lowmem, which is a waste of address space.  Most memory
allocations go to userspace, which doesn't care if the memory
came from lowmem or highmem, and allocates pages individually,
which will fragment lowmem.  Increase the vmalloc space to 768MB,
which will reduce lowmem to 128 MB.

Change-Id: I8c0940b1d3f235d518181e712d240bef7050ab08
Signed-off-by: Dan Murphy <dmurphy@ti.com>
Signed-off-by: Colin Cross <ccross@android.com>