arm: omap: Add always_on contraint flag for VDAC ldo
With this below commit:
2b337d7d1519eef62f9901de443629d6f110a043
[Tablet] board file: Set the regulator full contraints

The regulator framework will shut down any unused
regulators in the system. This was causing the VDAC
ldo to be shut down, which powers the clock generation
chip CDC3SO4(TCXO and driver) that supllies clock to
the ULPI Phy, thereby breaking EHCI enumeration.

This patch adds the always_on flag for VDAC ldo for
Blaze and Blaze Tablet.

Change-Id: Ie707a32c9d0a564bec4b96a3776cdc4b2c00ad57
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>