ASoC: SDP4430: Initialize the twl6040_codec struct
In commit b8169e9f, the sdp4430.c added a module-scope pointer to
struct snd_soc_codec *twl6040_codec, and initialized it with the
module (sdp4430_soc_init()).  However, when the code was merged into
the p-android-omap-3.0 branch (4de21ef0) the code became disabled
because it followed an unconditional `return 0;`.  This left the
pointer unitialized.

Therefore, the initialization code is moved up sooner in the init
function, and also checked for error.

Change-Id: Ia365e09ba8fccedb3b4432616b0e55b3ec7ea438
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>