[Tablet] omap4: Add dmm_init to the tablet board file
The power management system uses the order of device creation to
determine the order of suspend and resume function calls.  A number of
existing drivers are dependent on the Tiler resuming before they do.

The dmm and Tiler devices need to be created in the board init functions
so that they will be earlier in the device tree.

Change-Id: Ic9232d101fb4543b8cbe2a6ac734c184e00aba06
Signed-off-by: Dan Murphy <dmurphy@ti.com>