usb: musb: Correct the sequence for DMAENAB/DMAMODE for mode1 RX
The existing sequence was not correct since although you program
the dma in mode one, you clear DMAMODE and musb controller will
see it in mode0 which is not correct.

The correct sequence is to set DMAENAB, set DMAMODE and set
DMAENAB again, this is needed to to get DMAReq to activate.

Change-Id: I1a99984e8196c9ede432e831e40ce7dab021e066
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>