ASoC: OMAP: Switch DMIC clocks back to twl6040
The digital microphones (DMIC) are designed to be used with the pad
clocks, which are powered by the twl6040.  These have been powered by
the system clock because there was no mechanism in place to ensure
that the twl6040 is powered up.  This change ensures that the twl6040
is powered up and then configures the DMIC to use the pad clocks.

Change-Id: I1a31209a3d32c89501ab465c63ff0fe1945aff46
Signed-off-by: Gabriel M. Beddingfield <gabrbedd@ti.com>