omap4:emif: provide self-refresh timing.
Define timing for DDR to be in a safe value which
is recommended to be 16 DDR clk cycles.
For 400 Mhz the emif idle time before putting the memory
into selfrefresh is 5.12 us (2048 ddr cycles @ 400 MHz)
when the memory is at full speed. The lower idle time causes significant
performance issues.

Change-Id: If97873f56c94f33b39513a44cfd3c0cea26d7d7e
Signed-off-by: Rajeev Kulkarni <rajeevk@ti.com>