usb: musb: Hold a timed wakelock for usb-autosuspend in OTG mode
In OTG mode, With no device attched in the first resume path USB
core layer is enabling musb device (runtime_resume) and USB core
layer waits for 2 seconds (auto-suspend idle-timeout) and then
will disables the musb device (runtime_suspend) again. But before
it can get a chance to do so, if a system wide suspend happens
L3 fails to hit the target state.

To avoid this scenario, we hold a timed wake-lock for 3 seconds.

Change-Id: Ib8b5a25b76ca23c951000f205898dd39ed05c4f7
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>