omap-musb: Fix hot-unplug issue
The mentor core needs to be signaled via the SESSEND bit
of the USBOTGHS_CONTROL when USB cable is plugged out from
host PC. But the PHY clocks were getting cut as soon as
the SESSEND bit is being set and so the Mentor core
is not getting the disconnect interrupt.

Remove enable/disable the PHY clocks from the omap4430_phy_power()
function and move it to omap4_charger_detect(). Runtime
clock enable/disable for phy is taken care of
by the pm_runtime_put_autosuspend which will call the
otg_suspend in the omap2430_runtime_suspend function.

Change-Id: Id6540e198c69f14f3b7b920b51b700b5f980e31d
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>