Initial Open Color format Definitions (OCD).
OCD is a set of Open Color format Definitions.  Color formats
are used in everything from BLTers to video codecs,
and from cameras to displays.  But although there are
plenty of common formats, the definition of the code specifying
them is never the same.

OCD attempts to solve this problem.  It provides logical color
format codes, and is extensible.

Change-Id: I21efada119207773c10f9b199b7a7ca0f1aa60d2
Signed-off-by: Rick Tillery <rtillery@ti.com>
Signed-off-by: David Sin <davidsin@ti.com>