usb: musb: Dont start Host session on probe unless device attached
This patch makes sure that we dont start the Host session unless
there is a cold boot scenario and a device(mouse) is attached.

Without this patch, the Host session was started by default
causing the gadget configuration to be not recognized by
the Host machine on cold boot in device mode.

Change-Id: Ibe8609a77dcb2c5c7455062a04337e4c123ce8a6
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>