Fix for Wi-Fi ON/OFF under regression scenario causing Kernel Panic
Wifi enable/disable regression resulted in kernel crash.
In regression scenario it was seen that access to registers
were being done after call to disabling clocks while enable
clocks had not happened. Fix for this was to replace the API
with mmc_claim_host and mmc_release_host APIs which took
care of enabling/disabling clocks along with incrementing and
decrementing of counters.

Change-Id: Ia135678a6857d1f1ff6828c0654c9d5aeeb20383
Signed-off-by: Chittim, Madhusudan <c-madhu@ti.com>