ASoC: SDP4430: Ignore suspend for legacy DMic DAI link
Commit 52ba67b (ASoC: Force all DAPM contexts into the same bias
state) has a negative effect on card's bias level due to widgetless
contexts. Bias level for widgetless contexts is calculated based
on the DAPM event. For example, for RESUME event, it bumps the
card's bias level up to power-on state. This is suboptimal and will
cost additional power transitions to the sound card.

In the meantime the ASoC framework fixes that problem, the side
effect of it can be prevent by making the corresponding DAI link
(DMic legacy) as ignore-suspend capable, so that SUSPEND/RESUME
DAPM events are not sent because of DMic legacy DAI link.

Change-Id: I6c4e7d5c7d855b29213cfde402eaf4d72d3975d7
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>