ASoC: omap-mcasp: Relax S/PDIF clocks/frame constraint
The first step in the current McASP clock dividers calculation is
very strict with respect to the McASP FCLK being an integer multiple
of 128 (number of clocks required for each S/PDIF frame).

The clock divider calculation algorithm doesn't search for exact
matches, but for dividers which can produce clock drift within an
allowed PPM range. For this reason, rejecting McASP FCLK rates not
divisible by 128, reduces the space of rates the algorithm can work
on. By simply including the clocks per frame value in divider
calculation we give more flexibility to the algorithm.

Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>

porting from p-android-omap-3.0 to p-linux-omap-3.4 as part of k3.0 to k3.4
migration effort for OMAP4 Tablet2 support

Change-Id: I37cae44a9ad79ed43fde55eb1ff95e68507017a4
Signed-off-by: Ranganath Krishnan <ranganath@ti.com>