OMAP4: PM: omap4_trigger_ioctrl: use proper mask for proper delay
PRM_IO_PMCTRL register has more fields beyond WUCLK_STATUS.
Shifting the bits does not guarantee that the resultant value
represents the WUCLK_STATUS (a single bit). Instead, mask the
same to be sure about picking up just the register field we need.

Without this, we may not manage to get the test to pass and it
times out after 1ms in every attempt. This wrongly forces the
prcm_interrupt_handler to be delayed by 1ms when IOPAD wakeup
is triggered.

Change-Id: Ib0a052968796181b4b252afc13bca8f12cdd940b
Signed-off-by: Girish S Ghongdemath <girishsg@ti.com>