OMAP4: PM: omap4_trigger_ioctrl: use proper mask for proper delay
PRM_IO_PMCTRL register has more fields beyond WUCLK_STATUS.
Shifting the bits does not guarantee that the resultant value
represents the WUCLK_STATUS (a single bit). Instead, mask the
same to be sure about picking up just the register field we need.

Without this, we may not manage to get the test to pass and it
times out after 1+ms in every attempt. This wrongly forces the
prcm_interrupt_handler to be delayed by 1ms when IOPAD wakeup
is triggered. This causes the IO daisy chain blind window to
be abnormally long, instead of a few us.

Measurements show that time between rising and falling edge
of WUCLIN is:
* without fix: 1600us
* with fix: 1.2us

Also the latch is really complete when the WUCLKIN pulse goes
low, so we need to wait until WUCLK_STATUS goes low.

Change-Id: Ib0a052968796181b4b252afc13bca8f12cdd940b
[d-elaidi@ti.com: parallel fix]
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>
Signed-off-by: Girish S Ghongdemath <girishsg@ti.com>