OMAP4: Tablet: Implement PDM UL errata
When TWL6040 enters into Sleep/Deep-Sleep mode, the value of PDMUP could be
at logic 1 (VDDVIO) state. In this state, the pulldown resistor on OMAP side
results in a current drain.

Since the logic 1 reset value cannot be modified on the TWL6040, the solution
to stop the current drain is to disable the pulldown resistor on the input of
the PDM buffer of OMAP.

Change-Id: Iea8919082009503b9ecdc476723d5facb6fd88a1
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>