OMAP4: HSI: Check HSR mode before clock disable
Added a security check of HSR mode vs CAWAKE level and in case there is a
mismatch, take recovery actions if needed.
Also do not check HSR mode in 3-wires before disabling clocks because HSR
state machine shall not be set in SLEEP mode if HSI is in 3-wires mode.
This can cause link stuck.

Change-Id: I9dda46416e1e795d48341fc44448dca2b573cd5a
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>