omap: resmgr: add omap resources which does not have a driver
adding resources:
- IVA (Image and Video Acelerator)
- IVA_SEQ0
- IVA_SEQ1
- FDIF (Face Detection InterFace)
- SL2IF (Shared L2 InterFace)
- ISS (Image Sub System)

As all this devices have a associated device but they don't have a driver.
However, we only need basic turn on/off functionality which be done using
the pm_runtime framework which only requires the device pointer. So,
this patch is also adding a generic function which can request a resource
only using the device and pm_runtime framework and this function is used
for all these resources. This function only allows exclusive access to
resource, that means that these resources can be used for only one user.

Signed-off-by: Fernando Guzman Lugo <fernando.lugo@ti.com>