usb: musb: Turn off the PHY while doing interface selection
The OMAP4 TRM says that for USB OTG, interface selection
has to be done before the PHY is activated and is not allowed
to change when the PHY clock is already running.

For charger detection, the interface selection was done with
the PHY kept on, this was resulting in random hotplug failures
at times. This patch corrects the issue.

Change-Id: I2385c4e5d5a3d669452af4507557c158e7b48dfc
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>