ASoC: twl6040: Add wakelock for headset detection interrupt
TWL6040 is kept active during a voice call, but CPU can go into
suspend state. If headset is inserted or removed while the CPU
is in suspend, SYS_NIRQ2 IRQ line will wake the CPU up and the
headset state will be acknowledged at kernel level, but the
event propagated to userspace layers to cause audio rerouting
might not be completed on-time before CPU goes back to suspend.
A wakelock is added to prevent CPU to go back to suspend
immediately after the event, instead it'll wait for 2 secs for
the event to be processed in userspace.

Change-Id: I88a71345c7a082801a21335f20c920c68bacb8fc
Signed-off-by: Misael Lopez Cruz <misael.lopez@ti.com>