blaze/blaze tablet: Set MMC5 built_in flag
When WLAN driver is loaded, MMC_PM_KEEP_POWER flag is set/managed
by the SDIO function driver every time system suspends/resumes.

This change is to handle scenario when WLAN driver is not loaded.

If this flag is not set WLAN chip is powered off during suspend and
powered on during resume in WLAN driver unloaded case.

Change-Id: Ifececb0d031fdc05b9cd885bcb97428a171a685e
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>