OMAP4430: OPP/voltage: 1.2GHz needs different SR ntarget
nTarget value for 1.2GHz uses the nTarget efuse offset meant
for 1GHz. This can cause instabilities since the voltage levels
are lower than required for operating at 1.2GHz.

Since SR nTarget values are hooked to omap_volt_data structure and
Vnom for 1.2GHz is same as 1GHz, we need to ensure a uniquely
identifiable voltage for 1.2GHz to supply the right offset.
We do this by adding a 10mV step on top of 1GHz Vnom. This
additional step should have no impact as SmartReflex AVS will
adjust this to optimal voltage on enabling SR from Vnom in operation.

With the unique voltage entry in omap_volt_data for 1.2GHz, we
provide the right efuse offset for 1.2GHz and map the OPP entry
for 1.2GHz to use this "OPPNITROSB" voltage. This will allow
voltage and SmartReflex AVS mapping back to the right nTarget
offsets + adjust voltage to the correct optimal voltage levels
for 1.2GHz.

Change-Id: I1e8d0efa46cfb5d6db1b54ef5fc323d0b7877fef
Signed-off-by: Taras Kondratiuk <taras@ti.com>
Signed-off-by: Eugen Mandrenko <ievgen.mandrenko@ti.com>