OMAP4: HDMI: Fix HDMI_TXPHY_TX_CONTROL.FREQOUT programming
According to TRM, the HDMI_TXPHY_TX_CONTROL.FREQOUT bit has to be programmed
based on the clock frequency. But this bit is always programmed as 1 in the HDMI
driver irrespective of the clock frequency values. Fix this and program
HDMI_TXPHY_TX_CONTROL.FREQOUT based on the clock frequency.

Change-Id: Ice48cf7eff7c3e5b8fc365634d41e38e176f2513
Signed-off-by: TAKAHASHI, Masato <m_takahashi@ti.com>
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>