omap4: hdmi: adding a macro for acr wa dependency with rpmsg and ipu
Hdmi acr workaround is based on the creation of an rpmsg channel
to start/stop the workaround in the ipu remote processor. If the ipu
remoteproc or rpmsg are disabled by menu config, the build is broken.

This patch adds a macro for disabling hdmiwa in the case of rpmsg
or remoteproc_ipu are disabled.

Change-Id: I1618d886d16ffe0c07b810e6e1fd678debc547cd
Signed-off-by: Juan Gutierrez <jgutierrez@ti.com>