omaplfb: Enable blit buffers to use vram
omaplfb will be configurable to support blit buffers either in TILER2D or in
the reserved VRAM area.

There is no-coexistence support for swapchain buffers at this point with
dsscomp enabled, but the changes here should be ready for this scenario.

Blit framebuffers are given their own stride rather than use the stride
calculated for the swap chain, this makes sense in order to interchange
usage of TILER2D & VRAM.

It should be noted that blit framebuffers are additional buffers for the
same device being serviced by a display swapchain.

Change-Id: I544ef89ae9b02ce4726350938f088733ec914b15
Signed-off-by: Tony Lofthouse <a0741364@ti.com>