MUSB: Disable PING for control ep in OTG host mode
We are observing that USB 3.0 hub is not supporting PING request
from OTG host device as a result it fails to enumerate. Disable
PING packet in OTG Host mode for control ep and we dont see this
issue any more. We analyzed traffic between a USB2.0 PC host and
the USB3.0 device and observed that the host does not send PING
packets during enumeration. Turning off PING packets in the OTG
host for control ep resolves this issue. Although we observed
this on a USB 3.0 hub, the issue could be there with other
USB 2.0/3.0 devices if they dont support PING protocol

Ported from:
p-android-omap-3.4: 1aaa162637 [Huzefa Kankroliwala <huzefank@ti.com>]

Change-Id: I39bd700ca2faf1f2eb11b19217f27a4c47ef8a31
Signed-off-by: Huzefa Kankroliwala <huzefank@ti.com>
Signed-off-by: Ruchika Kharwar <ruchika@ti.com>
Signed-off-by: Alexander Savchenko <oleksandr.savchenko@ti.com>