OMAP4+: HSI: reduce RX interrupt number in polling mode
When using RX polling, the RX interrupt was never disabled once it was
triggered, causing extra RX interrupts to be raised again.

But when in polling mode, a single RX interrupt is enough to notify hsi_char.
So this fix disable the RX interrupt once it has triggered.
It is now clearly hsi_char (or equivalent Link Layer) responsibility to
re-enable polling once read is over.
This fix also explicitely disable RX interrupt when closing the channel.

Change-Id: Ic6e9b645f7d4530712fa2766ce99164a9020ebaa
Signed-off-by: Djamil Elaidi <d-elaidi@ti.com>