OMAPDSS : DSSCOMP : Avoid pushing content if no panel is active
Currently we push content on a blanked manager in case it is
default HDMI display. This causes a lot of logging in dsscomp call
back when we apply blank to manager.

Check added to see if number of overlay in the composition are not
zero.

In HWC we are checking if the default display is not present
we will populate the composition number of overlay as zero.

Dependencies:
This patch has dependency on HWC patch mentioned below..
http://review.omapzoom.org/#/c/26472/ - JB release

Change-Id: Idffd022e293949bfea260a01b96b8f6eff8327d3
Signed-off-by: Dandawate Saket <dsaket@ti.com>
Signed-off-by: Varadarajan, Charulatha <charu@ti.com>