OMAP4: hack for device is not resuming after suspend
The device is not resumed, in the case if "POWERON" button was
pressed for short time. The algorithm is following:
1. Device is resumed, and system reads the reason of resuming
from TWL status registers. But user releases the "POWERON"
button while system is in resume process. See pic1 and pic2
                          ___
   pic1                  |   |
real button state       /|\ \|/
                 ________|   |___________

   pic2                       |   |
system receive               \|/ \|/
following state  _____________|___|______

      |                            |
     /|\                          \|/
where | is button pressed state    | is button released state

2.We are read the cause of resume and get "button released".
3.After that we get second interrupt "Button released", because
button was really released and second interrupt was happened.
See pic2.
In this situation the system sees this, like two pairs
of "pressButton/releaseButton".
The patch removes a second "pressButton/releaseButton".

Change-Id: Ieec7227db19ed43f7964159169eee44386d7d511
Signed-off-by: Eugen Mandrenko <ievgen.mandrenko@ti.com>
Signed-off-by: Sergii Postulga <x0153364@ti.com>