Regulator: TWL6032: Keep ON some resources during suspend
For TWL6032 revision < ES1.1 with EEPROM revision < rev56.0
LDO6 and LDOLN must be always ON because of the hardware
bug in the TWL6032. If LDO6 or LDOLN is always on then SYSEN
must be always on.
For TWL6032 revision >= ES1.1 with EEPROM revision >= rev56.0
those LDOs can be off in sleep-mode.
This patch performs additional checks.

Change-Id: Ifb7d894e2f6282b04eef8f27e993e8235ad7efc1
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>
[oleksandr.tyshchenko@ti.com: ported from K3.0]
Signed-off-by: Oleksandr Tyshchenko <oleksandr.tyshchenko@ti.com>