OMAP4: Regulator: Correct LDOs state in STATE_MEM
All regulators have state_mem settings with flags 'enabled'
and 'disabled'. If both flags have the same value (true or
false) LDOs have incorrect setting and there are some
warnings before entering the device to the sleep-mode.
This patch correct settings for the LDO6, LDOLN and SYSEN
for the TWL6032.

Change-Id: I50b29ae503954826363279225b13967139106507
Signed-off-by: Oleksandr Dmytryshyn <oleksandr.dmytryshyn@ti.com>