gccore: fixed power off/on transition.
Simplify the power transition functionality, there is now just
an OFF and ON transition and the initial state is UNKNOWN.

This patch also enables GC320 to power off when at suspend time
(though it does not have an explicit suspend/resume handler).

ON means running at full speed, it enables the GC320 FCK, clears
the GC320 out0 idle state and enables interrupts.

OFF means enabling pulse keeping, setting the idle bit in the out0
GC320 status register and disabling the functional clock for GC320.

However note that this does not explicily power off the GC320 core,
however when the DSS & the system enters suspend the interface clock
appears to be disabled and this results in the GC320 h/w turning off.

Change-Id: I7d3187ca0b5dd803af81c22a8407067a6aaeb5cd
Signed-off-by: Alexei Shlychkov <shlychkov@gmail.com>
Signed-off-by: Tony Lofthouse <a0741364@ti.com>