tty: omap-serial: call pm_runtime_put/autosuspend in the error cases
In the error cases the runtime_put call is missed. This patch intends to fix the
same. In case dma request fails, we fall back to the nondma mode so after
enabling the threshold call put_autosuspend.
The autosuspend version is used to save unnecessary clock cut/enable
in quick succession. Also is inline with the stratagy taken in the non-dma
case.

Cc: stable@vger.kernel.org
Cc: Govindraj.R <govindraj.raja@ti.com>
Signed-off-by: Shubhrajyoti D <shubhrajyoti@ti.com>