gpu: omap_hwc: Adds a flag to hint omaplfb a blit FB is needed
When HWC generates blits a DSS pipe is configured and assigned
for the blit framebuffer, this patch adds a flag to give a
hint to omaplfb to use such pipe when it is present.

This helps in situations where there are not blits but blit
buffers to synchronize with and a blit buffer must be used.
In such case the rgz_items and num_ovls from dsscomp where not
enough data to recognize this scenario.

Change-Id: I2bcb13e6341189528f4ffd73c03badd431afbf15
Signed-off-by: Gustavo Diaz Prado <a0273371@ti.com>
Signed-off-by: Sundar Raman <sunds@ti.com>