OMAP: Board ID: add dump information to indicate which SoC the issue is seen
mach_panic_string was added with the intent for SoC and machine information
to be represented so that bug reports which contain kernel crashes have
Board and SoC information as well for aiding debug context.

This especially saves multiple rounds of debate between developers and
reporter especially when information is reviewed at a later point of time.

Example dump with this change:
[    5.295288] Internal error: : 1406 [#1] PREEMPT SMP
[    5.336639] Modules linked in:
[    5.336639] CPU: 1    Tainted: G        W    (3.1.0-00021-gf6c339c-dirty #14)
[    5.336669] PC is at omap_i2c_wait_for_bb+0x90/0xb4
[    5.336700] LR is at omap_i2c_wait_for_bb+0x18/0xb4
[    5.336700] pc : [<c03733a0>]    lr : [<c0373328>]    psr: 40000013
<snip>
[    5.339416] Backtrace:
[    5.339447] [<c0373310>] (omap_i2c_wait_for_bb+0x0/0xb4) from [<c0373944>] (omap_i2c_xfer+0x6c/0x518)
<snip>
[    5.339721]  r7:00000013 r6:c004aa04 r5:c00634ac r4:e9847ba0
[    5.339752] Code: e1a03311 e19230b3 e3130a01 1afffff1 (e3a00000)
[    5.339889] Board Information:
[    5.339889]  Revision : 0000
[    5.339889]  Serial	: 0000000000000000
[    5.339919] SoC Information:
[    3.339919] CPU	: OMAP5430
[    5.339919]  Rev	: ES1.0
[    5.339919]  Type	: HS
[    5.339919]  Production ID: D0110451-2E5080CC
[    5.339919]  Die ID	: 7DB60020-00000809-015457CE-0800F00D
[    5.339919]
[    5.339935] ---[ end trace 1b75b31a2719ed1e ]---

At this point, we don't use the machine_name info which is restricted inside
ARM kernel core code, so it could be exported OR a new API could be exposed
for boards to register their names explicitly in the future.

Change-Id: I6e82adde37458fe21ae578f86e4f8939a37cb8e0
Signed-off-by: Nishanth Menon <nm@ti.com>