ARM: OMAP5: hsmmc: evm: declare support for MMC_PM_KEEP_POWER
Declare support for keeping the power of the wlan chip
while suspended. This is needed for Wakeup-On-Wireless.

Declaring this capability on mmc host has two implications
1) This will enable Wake on Wirelss WLAN feature in mac80211 stack
2) This will prevent WLAN chip from being shut down during host suspend.

This has some power impact on overall system power as WLAN chip will be on.

Android needs this feature of staying connected during suspend.
User has the option to disable this in Android settings if they don't
want to stay connected.

Change-Id: I421046d469c10dd66729b03bf44371aee8720e5b
Signed-off-by: Viswanath Puttagunta <vishp@ti.com>