ARM: OMAP2+: PM QoS: control the power domains next state from the constraints
When a PM QoS device latency constraint is requested or removed the
constraint is stored in the constraints list of the corresponding power
domain, then the aggregated constraint value is applied by programming
the next power state using omap_set_pwrdm_state.

Tested on OMAP3 Beagleboard and OMAP4 Pandaboard in RET/OFF using
wake-up latency constraints on MPU, CORE and PER.

To make the search more accurate an reverse lookup table is used
to covert from search index to power domain state to check for.

[grygorii.strashko@ti.com: fixes]
Signed-off-by: Grygorii Strashko <grygorii.strashko@ti.com>
Signed-off-by: Jean Pihet <j-pihet@ti.com>