gccore: Remove Android early suspend late resume handlers
Early suspend/late resume handlers are primarily present for the purpose of getting
the screen into a low power state for Android.

On the OMAP platform the principle means of handling this is that DSSCOMP will
blank the screen.

In order to enter the lowest power state gccore should use the standard Linux
suspend/resume handlers.

Change-Id: I3dded7c5c83bc2f03048d6c8ef32fce697ab847c
Signed-off-by: Tony Lofthouse <a0741364@ti.com>