TI thermal framework: improve thermal domain manipulations
This patch changes the thermal device data structures
to have a pointer to thermal domain. This way, once
an operation is required on a thermal device, there is
no need to iterate on thermal domains lists.

This patch also centralizes the operations to search and create
thermal domain data structures. Idea is to improve code readability,
having these operations done in one place only.

The patch does not change the code functionality so far.

Change-Id: Ibc49a40d9bfce1710fd790317dfa026c74651bd3
Signed-off-by: Eduardo Valentin <eduardo.valentin@ti.com>
Signed-off-by: Radhesh Fadnis <radhesh.fadnis@ti.com>