ARM: OMAP4: Reduce the static IO mapping
EMIF, GMPC and DMM driver can ioremap() the address
space as part of driver intialisation and there is
no need to have static IO mapping for them.

Hence remove the un-used static IP space and let
the respective drivers manage it as part if driver
init.

Signed-off-by: Santosh Shilimkar <santosh.shilimkar@ti.com>
minor:
pull to p-linux-omap-3.4

Change-Id: Ibb7487f299395ad931c78bd8d0fe42641a89664f
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>