ARM: OMAP2+: irq: Add support for more no of interrupts
With addition to TI81XX, AM33XX family of devices, the number
of interrupts supported has increased to 128, compared to 96
in OMAP3 family of devices.
The current implementation for irq handling is hardcoded to
96 interrupts (with 3 registers to handle), this patch cleanups
the code, to increase maximum number of interrupts support
to 128, with dynamic detection of no of registers required for
handling all interrupts.

NOTE: Ideally, we should use dynamic allocation to allocate memory
      for registers, will align with community and implement/submit it.

Signed-off-by: Vaibhav Hiremath <hvaibhav@ti.com>
Signed-off-by: Afzal Mohammed <afzal@ti.com>